@@ -16,6 +16,7 @@
     "classnames": "^2.3.1",
     "framer-motion": "^6.2.8",
     "node-sass": "^7.0.1",
+    "prettier": "^2.6.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.28.1",
@@ -23,20 +24,16 @@
     "react-redux": "^7.2.6",
     "react-scripts": "5.0.0",
     "redux-persist": "^6.0.0",
+    "typescript": "^4.6.2",
     "use-match-media-hook": "^1.0.1",
-    "web-vitals": "^2.1.4",
-    "yup": "^0.32.11"
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "lint": "eslint --ext .ts,.tsx src --fix",
+    "lint": "eslint */**.{tsx,ts}",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
-  },
-  "prettier": {
-    "singleQuote": true,
-    "printWidth": 100
   },
   "eslintConfig": {
     "extends": [
@@ -57,15 +54,13 @@
     ]
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.15.0",
-    "@typescript-eslint/parser": "^5.15.0",
-    "eslint": "^8.11.0",
+    "@typescript-eslint/eslint-plugin": "^5.18.0",
+    "@typescript-eslint/parser": "^5.18.0",
+    "babel-eslint": "^10.1.0",
+    "eslint": "^8.12.0",
     "eslint-config-google": "^0.14.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.29.4",
-    "prettier": "^2.6.0",
-    "typescript": "^4.6.2"
+    "eslint-plugin-react": "^7.29.4"
   }
 }
