@@ -1,9 +1,9 @@
-import { FC } from 'react';
+import React, { FC } from 'react';
 import { footer } from './data';
 
 import { ILinks } from './types';
 
-const s = require("./Footer.module.scss") as any;
+import s from './Footer.module.scss';
 import { ReactComponent as Logo } from '../../img/logo.svg';
 
 
