@@ -1,6 +1,11 @@
-declare module "\*.svg" {  
+declare module "*.svg" {  
     import React = require("react");  
     export const ReactComponent: React.SFC<React.SVGProps<SVGSVGElement>>;  
     const src: string;  
     export default src;
+}
+
+declare module '*.module.scss' {
+    const classes: { [key: string]: string };
+    export default classes;
 }
